pyproject.toml
[project]
name = "ai_assistant"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "@gilberto23", email = "gilberto@helicon.ai" }
]
dependencies = [
    "openai>=1.53.0",
    "langchain-community>=0.3.4",
    "litellm==1.51.0",
    "langchain-ollama>=0.2.0",
    "autogen-agentchat~=0.2",
    "ipykernel>=6.29.5",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "torchaudio>=2.5.1",
    "matplotlib>=3.9.2",
    "deep-phonemizer>=0.0.19",
    "chromadb>=0.5.17",
    "markdownify>=0.13.1",
    "pypdf>=5.1.0",
    "transformers>=4.46.1",
    "sentence-transformers>=3.2.1",
    "langchain-openai>=0.2.5",
    "faiss-cpu>=1.9.0",
    "langgraph>=0.2.45",
    "dspy>=2.5.29",
    "dspy-ai==0.1.9",
    "langserve>=0.3.0",
    "sse-starlette>=2.1.3",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ai_assistant"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff]
line-length = 88
show-fixes = true
select = [
    "F",   # Pyflakes
    "W",   # pycodestyle
    "E",   # pycodestyle
    "I",   # isort
    "UP",  # pyupgrade
    "PL",  # Pylint
    "T201", # Print Statement
]
ignore = ["E501"]  # Ruff format takes care of line-too-long


